app/backend/requirements.txt
spacy
sentence_transformers
openai
flask
flask_cors
firebase_admin
google.cloud.firestore